pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hapnet"
version = "0.1.0"
description = "Population-aware haplotype network analysis and visualization in Python."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
  { name = "Andrew A. Davinack" }
]

dependencies = [
  "biopython>=1.80",
  "networkx>=3.0",
  "matplotlib>=3.7",
  "numpy>=1.24",
  "scipy>=1.9",
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[project.urls]
Homepage = "https://github.com/parasiteguy/hapnet"
Repository = "https://github.com/parasiteguy/hapnet"
Issues = "https://github.com/parasiteguy/hapnet/issues"

[project.scripts]
hapnet = "hapnet.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
